+++ scipy/interpolate/__init__.py
@@ -141,6 +141,7 @@ Additional tools
 
    lagrange
    approximate_taylor_polynomial
+   pade
 
 .. seealso::
 
@@ -187,6 +188,8 @@ from ._cubic import *
 from .ndgriddata import *
 
 from ._bsplines import *
+
+from ._pade import *
 
 __all__ = [s for s in dir() if not s.startswith('_')]
 from numpy.testing import Tester
